services/worker/requirements.txt
watchdog>=4.0
pymupdf>=1.24
python-docx>=1.1
docling>=2.0
httpx>=0.27
chromadb==0.6.3
tiktoken>=0.7
